requirements.txt
-e .
aws-cdk.core
aws-cdk.aws-eks
aws-cdk.aws-ec2